pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "p2obp"
version = "1.4.2"
description = "The Phase 2 OB Pipeline (p2obp) an automation tool for the MATISSE observation preparation."
requires-python = ">=3.9"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["astronomy", "observations", "observation-preparation"]
authors = [
	{name = "Marten Scheuck", email = "martenscheuck@gmail.com"}
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
	"Framework :: Pytest",
	"Framework :: Sphinx",
	"License :: OSI Approved :: MIT License",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"Topic :: Scientific/Engineering :: Astronomy",
	"Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
  "importlib-metadata >= 5.0.0; python_version < '3.8'",
  "astropy >= 5.1.1",
  "astroquery >= 0.4.6",
  "attrs >= 22.1.0",
  "numpy >= 1.23.4",
  "pytest >= 7.2.0",
  "toml >= 0.10.2",
  "p2api >= 1.0.5",
  "openpyxl >= 3.1.2",
  "pyephem==9.99",
  "pandas >= 2.0.1"
]

[options.optional-dependencies]
dev = [
    "ipython >= 8.12.1",
    "pytest >= 7.2.1",
    "pytest-cov >= 4.0.0",
    "sphinx == 7.0.0",
    "sphinx-autobuild == 2021.3.14",
    "sphinx-autodoc-typehints == 1.22",
    "pydata-sphinx-theme==0.13.3",
    "myst-parser == 0.18.1",
    "numpydoc==1.5.0",
]

[tool.setuptools]
py-modules = [] # Removes the top-level package bug
include-package-data = true

[project.urls]
repository = "https://github.com/MBSck/p2obp"
